contrib: set pkg_config_libdir in meson crossfile
pkg_config_libdir is not a builtin option, to use
it we need to define it either in the meson.build of the
project or in the crossfile.

This used to fail silently, but since PR#8974 (meson 0.60.0.rc1)
it is a fatal error.

(cherry picked from commit 791c4833683a19bcf5de2ce946c07425224c901d)