resource: fix double vout stop on error
vout_Request() can be called on a started or stopped vout, it will try
to let the vout alive even in case of error (to let the user decide what
to do). There are some corner case where it returns an error and the
vout is stopped. In that case, the input resource was trying to stop the
already stopped vout.

To fix this issue, also return the vout state from vout_Request().

Fixes #26345