qt: pass the updated item along with the new data to the thumbnailUpdated cb
Note that models that use vlc_ml_media_generate_thumbnail loose the notification
at this very commit, this will be reinstated in the following commits.

The former form of the signal was a bit inconsistent as the MLVideo item (for
instance) was updating itself (in an unsafe manner) with its own callback but
the model was notified through this mechanism. This could have potentially led
to data inconsistency.