qt: remove duplicate write
`data` is the pointer to an object belonging to the current category node.

when handling a 'general' subcat, which gets "merged" into the category
node, all we need to do is modify the `data` object's attributes. we do not
need to save the pointer to this object in the tree node object, which was
already done earlier.