qt: rework adv prefs tree to use derived `QTreeWidgetItem`
rather than allocating a custom object to hold data for each tree node,
linked to the tree via `QObject` for ownership and destruction purposes,
and the pointer held in a 'data' property of the node, let's simplify this
significantly with a custom `QTreeWidgetItem`.

`PrefsTreeData` has been renamed `PrefsTreeItem` and converted as such.