qt: cache and reuse pointer to main module in prefs tree creation
each creation of a panel for a non-plugin (cat or subcat) node resulted in
a call to `module_get_main()` to obtain the pointer to the main module to
thus be in a position to get to the main config set.

by doing this once when creating the `PrefsTree` instance and caching it,
we can copy it into the existing module pointer property of cat and subcat
nodes (which was only used previously for plugin type nodes) when they are
created, and thus we will already have the necessary pointer available when
creating the panels, saving dozens of calls.