window: remove configuration parameter from Enable()
This removes the window configuration parameter from
vout_window_Enable(). The initial configuration is now set when the
window is created, and updated with the regular functions:
vout_window_SetSize(), vout_window_SetFullScreen() and
vout_window_UnsetFullScreen().

This reconciles the window configuration with the full-screen output
selection: if the window provider supports output selection, it can
simply ignore the full-screen flag in the configuration.

In practice, this makes no differences in the GL and splitter cases,
which were creating and enabling the window together. In the vout
case, this actually enables some small simplifications.

This change does not affect the module interface: the window
configuration is still passed via the enable() callback. However window
provider can choose to ignore all parameters other than the decoration
flag. Dimensions is already provided by the resize callback, and
full-screen status via set_fullscreen and unset_fullscreen callbacks.

It should be possible to remove the dimensions and full-screen flag
from the window configuration completely, but that is left for further
work, as it would affect many different modules.