core: add the vlc_ancillary API
Ancillaries can be attached to any vlc_frame_t or picture_t (in next
commits).

Ancillaries can be created from:
 - packetized demuxer modules
 - packetizer modules
 - decoder modules

It can't be created from a non packetized demuxer module since the
attachement to the vlc_frame will be lost by the packetizer module that
will be automatically inserted.

Ancillaries are automatically forwarded from a vlc_frame_t to an other
vlc_frame_t and from a picture_t to an other picture_t. This allow to keep
ancillaries untouched when audio filters or video filters are used (these
filters don't have to know about the ancillary).

Ancillary readers can be either:
 - A decoder module
 - An audio output
 - A video output
 - Any filters

Refs #18762