vout: always keep the original video format
This retains the original video format after the display is closed. The
main motivation for this is that the window needs a format to determine
its size, and to translate mouse coordinates, both of which can happen
after the display has been stopped.

Even though the format will be stale, this resolves several data races
in a simpler and saner way than checking if the display exists (which
requires the display lock).