vout: update window size regardless of display
As of the previous change, we always have an original format except at
the very beginning before the first attempt to start a display.

This allows the window size to be updated when the display gets started
in a vout that has an already enabled window. This also avoids taking
the display lock whenever compute the window size.

This essentially reverts commit 3784e7066210441174f1937151b60d3ae3bbb80a.