mp4: fix invalid use of undefined pointer state
if the `asprintf()` fails then the pointer is left in an undefined state.

should it fail and the append-to-list and any subsequent loops have managed
to make it through the low-memory condition successfully, then firstly some
code was inevitably going to read the pointer in trying to use the string,
and secondly at some point later when the list of seekpoints is destroyed,
`vlc_seekpoint_Delete()` would be calling `free()` on it.

the original version of this patch simply reset the pointer to null upon
failure - if anything is not designed to cope with a null name then that's
an entirely separate existing bug. however, two other places within this
file that make seekpoints, including the very similar `LoadChapterGpac()`
function, take the action of destroying seekpoints for which they have
failed to allocate the name, and feedback from review requested that same
behaviour here.