musicbrainz: fix wrong url
inconsistent with that used in `coverartarchive_make_releasegroup_arturl()`.

i believe that with-dash is likely the correct form, since firstly, that
function is actually used unlike this one, and this function was also
fundamentally broken before the previous commit. also secondly, the info
at [1] seems to indicate that with-dash is correct.

[1]: https://github.com/Borewit/musicbrainz-api