freetype: fix another leak on error
the `i_count` attribute of `p_text_block` tracks the "used" length of three
array attributes, and is only updated at the end of the function, upon
success.

if construction of the `p_rubyblock` object fails and the error path is
taken, the allocated style object `p_mgstyle` will have had its pointer
recorded in the `pp_styles` array, but since under the error path `i_count`
will not have been increased, the code later used to destroy the text block
object will ignore the portion of the realloc'd `pp_styles` array that
recorded that `p_mgstyle` pointer, and thus this object instance will not
be free'd by it creating a leak. furthermore if the function is called
again then the pointer would get overwritten by a new one.

we must thus ensure that we destroy it in the error path.