aom: use choice list for --sout-aom-profile
and add longtext (taken from the comments in the switch block).

this gives better UX, with a drop list in gui prefs rather than a
numberic control, a mapping of values in help output, and the
longtext provides additional details to further explain the
differences.

note that it is correct to pair a range with a choice list, since
only ranges are enforced upon setting values.