aom: update range of --sout-aom-usage
the author of this module seems to have aligned the usage of this option
with the 'usage' param of `aom_codec_enc_config_default()`, which takes
one of the following three values:

 - AOM_USAGE_GOOD_QUALITY=0
 - AOM_USAGE_REALTIME=1
 - AOM_USAGE_ALL_INTRA=2

the 3rd is a recent addition from libaom v3.1.0.

i don't see any harm in expanding this to allow this additional new
value to be passed through to libaom.