aom: update default for --sout-aom-lag-in-frames
the options seem to relate strongly to options provided by the aomenc tool,
and their current documentation for this option states that the default
they currently use is 19 ([1]). i don't see any harm in adjusting ours to
match.

---

note, i'm curious about the fact that our max range is 70, while their
documentation states 35, though there is conflicting information later on
in their documentation.

the author of our module has commented `MAX_LAG_BUFFERS + MAX_LAP_BUFFERS`
besides the range, which correspods to this in ([2]), but doesn't help me
understand at all:

```
 #define MAX_LAG_BUFFERS 35
 #define MAX_LAP_BUFFERS 35
 #define MAX_TOTAL_BUFFERS (MAX_LAG_BUFFERS + MAX_LAP_BUFFERS)
```

is aomenc maybe taking the user value (0-35) and adding `MAX_LAP_BUFFERS`
(35) to it, whilst we are expecting the user to account for this??

[1]: https://github.com/master-of-zen/Av1an/blob/master/docs/Encoders/aomenc.md
[2]: https://aomedia.googlesource.com/aom/+/refs/heads/applejack/av1/encoder/lookahead.h