opengl: vaapi: load GL functions from interop
Only glBindTexture is used from the interop code, so remove the
dependency from the vtable in vlc_gl_api_t altogether.

The rationale for removing vlc_gl_api_t is that interop implementations
are already required to load specific function for most of them (like
the eglCreateImageKHR function) and having vlc_gl_api_t in the public
module API require exposing the OpenGL vtable and types to the public
scope, which is pretty ard to justify and port/maintain for every
platform.