opengl: interop: remove vlc_gl_api_t references
The implementations are now independant of the vlc_gl_api_t structure
and the OpenGL virtual table.

The rationale for removing vlc_gl_api_t is that interop implementations
are already required to load specific function for most of them (like
the eglCreateImageKHR function) and having vlc_gl_api_t in the public
module API require exposing the OpenGL vtable and types to the public
scope, which is pretty ard to justify and port/maintain for every
platform.