opengl: fix missing option subcat
and a couple of syntax issues.

the hidden subcat is used here rather than `SUBCAT_VIDEO_VFILTER` because
according to discussion in !1181 this option is just used as a hack, not
meant for use as a normal user config option.

lack of a subcat entry when there are options is normally because the
author forgot it, which causes a bug of the options not being visible in
GUI preferences. we should at least have a comment here to highlight that
it is deliberate in this special case.

however, a test for option set validity could be added in future, which
would only end up failing on lack of subcat entry here, so to avoid that
let's add one. using the hidden subcat will "properly" keep the option
hidden from GUI prefs (and also potentially help output if the help output
logic ends up gaining the ability to also skip such hidden options).