display: partially inline vout_display_GetDefaultDisplaySize()
Only from vout_display_New() can (and should) the window size be used
rather than the ideal default display size. In the other call sites,
we actually want to compute the ideal size regardless of the actual
window size.

Take that case out of vout_display_GetDefaultDisplaySize() to enable
further refactoring. This is a non-functional change, except for the
added run-time warning.