display: ignore display size in default computation
In practice, this function is used to compensate mismatches between
the source format SAR and the display PAR and should only be based on
those parameters and the source format visible dimensions. Indeed, in
the two call sites of the function pass 0x0 as the display size.

In the vout_display_New() call site, the function is used to
work-around semi-broken window providers that do not provision a window
size, or only do so with a delay. Normally, it just copies the dimensions
from window_props.

Either way cfg->display dimensions are always zeroes.

Furthermore, the 3 hereto unused and hereby removed cases failed to
account for orientation correctly: from_source should have been false.