vout: fix palette double free
As of the last rework, ownership of the original format has already been
passed to the video output by the point of enabling the window. It is
kept until the next format change, or the destruction of the video
output. So it shall not be freed if enabling the window fails.