display: try next module if conversion fails
After a display module initialises succesfully, a conversion filter
chain must be built from the source format to the display format.
Currently, this is effectively assumed to succeed. When it fails,
the display module is closed and display creation is given up entirely.

However most display plugins accept practically any source format that
is thrown at them, with little or no consideration for the possibility
of conversion chain failure.

After closing the unusable display module, with this change, the next
display module in decreasing priority is tried.