transcode: --sout-transcode-?enc default to "none"
This picks a default value that actually belongs in the choice list. In
this particular cases, the default is not to transcode ES in the given
category. Note that this does **not** interfere with the
--sout-transcode-?codec options which are handled separately.

This also enables to distinguish between default and error, though this
change does not actually implement this.