splitter: have default --video-splitter to "none"
The choice list always offers "any" and "none", and in this case, the
default is to not have a video splitter. This change fixes a bug whereby
the GUI would save the value "none", which the core would then treat as
a module name. No module by that name exists, so this would ultimately
cause the video output to fail entirely.

Also this treats NULL as an actual memory allocation error, instead of
the default value which was previously undistiguishable.