macos: add a separate alternative macro for non-translation marker use
essentially the macos side of 6cd08480f5c3c0d7505c17c1a77ef5668c47c9a2,
as requested in !1172 discussion.

the new `NSTR()` macro is the equivalent of `qfut()`. it is intended for
use instead of `_NS()` when wrapping a define name, variable or function
call, where the system that scans for strings we want to be translated
should not apply, leaving `_NS()` to just be used for marking such strings.

i have avoided renaming `_NS()` to avoid unnecessary widespread change.