Revert "opengl: vout_helper: use pl_gpu_dummy for shaders"
This reverts commit 2437484d7684a4661048df1e095484a276bbea0b.

The referenced commit is not a good way of fixing the issue described in
its commit message. Attaching a dummy GPU to the pl_shader makes the
shader think a GPU is actually available, and can lead to e.g. the use
of a dither texture (which VLC has no mapping code for, and would thus
result in a shader compilation error).

All of the queried limits are also irrelevant for the way we use shaders
in this code, since they only affect GPU resources (which we can't use)
or texture sampling (which we don't use).

Fixes #26503 by removing the conflicting code