cmdline: re-implement the preliminary commandline argument processing
Processing of commandline arguments is done in two passes. The first is
done with only knowledge of the core option set, to pick up any use of a
few options that affect plugin data loading (`--plugins-cache`,
`--plugins-scan` and `--reset-plugins-cache`). After that is done then we
can load the plugin data, thus gaining knowledge of the full set of
available options, putting us in position to do the second pass, which
fully and properly processes everything.

The way the first pass was implemented was to run the normal processing
function with only the core option set loaded and with error/warning
reporting disabled.

This design was not ideal for two reasons:

 1. It was unnecessarily inefficient:
   - It involved creating a copy of the argument set to operate upon such
     that the original was left intact for the second pass.
   - It meant that we processed the core option set twice in creating a
     representation of options to give to getopt (this was not cached and
     re-used).
   - It meant that every argument was compared with a set of something like
     630 options in that first pass when we're only interested in looking
     out for use of 9 (3 variants per bool option). (Note that vlc binaries
     typically add a few arguments to the user provided set so there
     typically are some arguments being processed upon every execution).
   - It meant that for every core option matched we were executing the
     `var_Create()`, `var_SetFoo()` and for ints/floats also `var_Change()`
     functions twice over (once per pass) to store the corresponding state.

 2. It was a problematic design. For certain valid and invalid use of
    arguments, the results of the two passes can be very different. The
    results from the first were not inline with what you would normally
    expect from getopt parsing behaviour, and created a significant
    undesirable behavioural difference under certain conditions between
    core options and plugin options. In general you'd ultimately end up
    with a mishmash of the two results. (The results of the second merged
    over the top of the first, but not necessarily fully overwriting
    everything). (Please read the documentation of the new
    `config_CmdLineEarlyScan()` function to better understand this aspect).

To demonstrate the correctness/consistency issue, compare the results of
`vlc --gain --help` with `vlc --sepia-intensity --help`. In the former case
`--help` is consumed as the value for `--gain`, as it should be per getopt
parsing rules (despite how illogical this may seem - read the documentation
mentioned above that talks about this). In the latter example, you see the
help output. why? Because in the second pass it got similarly eaten up, but
in the first pass the `--sepia-intensity` option was unknown and so skipped
over, letting `--help` be picked up; the two results got merged, letting
`--help` incorrectly be applied (which is technically wrong behaviour).

This commit replaces the first-pass with a new function that very simply
scans over the argument set looking out for use of just those three very
specific options. This is vastly more efficient relatively speaking, and
solves the "correct" and "consistent" parsing issue as best as we can
without dropping the need for a preliminary scan.