core: move saved settings handling to after full cmdline arg processing
Switching the order here is necessary to prepare for reworking the
preliminary commandline argument parsing, since the intention is for that
to now explicitly just check for the options that affect plugin loading,
and thus use of core options regarding the saved settings file will not
have been captured yet at the point this was being done.

Additionally, it is simply more logical to do it after we've *completely*
parsed the cmdline args. Loading the saved settings will have been a
complete waste of effort should an error condition occur in cmdline
parsing.

Note that this in no way affects how CLI option use is prioritised over
saved settings for runtime behaviour.