core: improve some init stage comments
The old comments against the actions of (1) loading of saved-settings, and
(2) the second-stage of commandline argument processing, made it sound like
firstly saved settings were loaded, then that state was overwritten by
state from use of CLI options, and thus the order of these two actions
needed to be in this particular order. In truth the two sets of config
state are stored separately and their order makes no difference to CLI args
taking priority in runtime behaviour.

The comment against use of `module_InitBank()` was incorrect since we don't
output core-only help early.

The comment against the first commandline argument processing stage (the
first use of `config_LoadCmdLine()`) was misleading, since in fact it does
not just "get" use of "options that affect module loading", but rather it
runs the processing code in a mode whereby error/warning output is disabled
and with it only having knowledge of the core option set, which results in
it broadly "getting" use of any and all core options (though this clumsy
early processing is imperfect).