contrib: do not forcefully add /usr/share
This was supposed to prevent native .pc files of the build system being
taken into use. But all this really did was add a directory to the
search path.

pkg-config environment variables are a little bit confusing:
- PKG_CONFIG_PATH directories are searched first.
  By default, it is treated as empty.
- PKG_CONFIG_LIBDIR directories are searched second.
  By default, the system native directories are searched, for instance
  /usr/lib/$(BUILD)/pkgconfig and /usr/share/pkgconfig.

If we want to search in /usr/share/pkgconfig, then it should be added to
LIBDIR where it would normally be. PATH is better suited for the contrib
installation directory.