contrib: medialibrary: link with -latomic
medialibrary uses atomics which may require libatomic depending on the
ABI. This is currently the case on Linux RISC-V. In principles, it
should be pulled as needed by `-pthread` which should be in LDFLAGS.
However:
1) The VLC build system passes `-pthread` in LIBADD instead of LDFLAGS.
2) GNU/libtool discards `-pthread` for CXX-tagged binaries.

This is essentially the same problem as with libupnp lacking
`-lpthread`. In this particular case, medialibrary.pc forces `-lpthread`
explicitly but lacks `-latomic`.

Fixes #26521.