configure: configure the native build of protoc with HOST name in contribs
Similar to how we force detection of luac there.

We might be using
AC_CHECK_TOOL(PROTOC, protoc, no, [${CONTRIB_DIR}/../bin$PATH_SEPARATOR$PATH])

Which works, but it doesn't keep the full path, so if the PATH is not set in
the environment when using HOST-protoc it doesn't find it. So we either have to
force everyone to add a new item in their PATH or just use the transparent way
that luac is using.