interop: vdpau: provide context in call macro
Provide the context in which the vdpau and OpenGL functions can be found
to ensure we are using initialized pointers. In particular, it now uses
the sys before it's set to the interop->priv part, instead of
dereferencing NULL.

Regression from b42d8e9efca53fe5f8e8542aaa2d6c1302674794.

Fix #26553

Co-authored-by: François Cartegnie <fcvlcdev@free.fr>