egl: allow swapping without current context
The vlc_gl_t API doesn't require the context to be current to call
vlc_gl_Swap(), but EGL needs it. Allowing vlc_gl_Swap from being called
outside of vlc_gl_MakeCurrent/ReleaseCurrent will allow to remove the
MakeCurrent/ReleaseCurrent mechanism and have the OpenGL provider pull
for a frame rendering instead of the code waiting on draw commands for
OpenGL to be ready to render, which is functionnally the same, but will
allow implementations like macosx.m or caopengllayer.m to be a real
OpenGL implementation instead of dummy provider around a display module.