lib: only change device of current output
Historically this function would set the device for a specified audio
output module. This required the application to somehow "know" what
audio output module it wanted to use, which is very suboptimal for
forward compatibility and portability.

Worse yet, this design failed outright in VLC 2.2 with the addition of
MMDevice and PulseAudio. This lead to the convention of leaving the
module parameter NULL to change the device of the currently active
audio output.

As with the removal of libvlc_audio_output_device_list_get(), it seems
saner to simply remove the old broken semantics.