test: use reproducible input
vlc_stream_NewURL() can spawn stream filters which may alter the content
or even the nature of the stream, which is not suitable here.

vlc_access_NewMRL() would avoid loading stream filters at all. But it
would also skip testing the cache filter, thus missing much of the point
of the test case.

So use a reproducible pseudorandom input using a constant seed, which
is known not to trigger any stream filter.

Fixes #26569.